requirements.txt
typer>=0.9.0
rich>=13.0.0
questionary>=2.0.0
pandas>=2.0.0
loguru>=0.7.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dateutil>=2.8.0
PyYAML>=6.0
pytest>=7.0